[swd-e2e-tests] Try to first flash with product bundle
This teaches the OTA tests how to download a product bundle, and then
attempts to use that product bundle to flash a device. Unfortunately we
still need to support flashing builds as old as F11 though, which don't
support PBv2, so we will fall back to the flash manifest if we can't get
the product bundle.

Change-Id: Ie3f7adb92caff89b5aac3841198f8bbe07ae9ae3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1090792
Reviewed-by: Mahesh Saripalli <maheshsr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Erick Tryzelaar <etryzelaar@google.com>