[graphics] Fix type of PFN_vkInitializeOpenInNamespaceCallbackAddr
This function was returning pointers for no real reason, so fix the
type. The return value was completely unused, so this won't affect
anything except UBSAN; UBSAN is disabled in this patch until the
follow-up that fixes it in the Vulkan-Loader repo.

Bug: 356893475
Change-Id: I07bee16e983f919eb95a09fc58982a0f01556fc9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1092874
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: John Bauman <jbauman@google.com>
Reviewed-by: Craig Stout <cstout@google.com>