[kernel] Remove last use of boot_reserve_foreach()
This was used when mapping EFI regions to double-check that they are
reserved (i.e., not marked as RAM). Rather than look for some PMM
replacement for this logic for this (currently) single callsite, it
seems more prudent to find an alternate, specific means of doing this
for now. In particular, this is straightforwardly done in this change by
consulting the normalized hand-off memory description and checking for
overlaps with RAM.

Bug: 347766366
Change-Id: Iaf88ca91d1a96a9df5bf8309aee578ffeabd7aee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093074
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>