[kernel][arm] MappingCursor for Unmap
Changes the UnmapPageTable method to use a MappingCursor to track the
unmap range, instead of overloading the return value, and adds
additional return information to track whether any entries were
unmapped or not, allowing the page_table_is_clear check to be skipped
when nothing was actually unmapped.

This change is almost universally positive in microbenchmarks and
improves scenarios by up to 15%. A handful of benchmarks that involve
unmapping large amounts of present page table entries show a regression
of up to 25%. This regression stem from the additional DEBUG_ASSERTs
that become present when performing the operations such as
MappingCursor::ConsumeVAddr.

Bug: 353386409
Run-All-Tests: True
Change-Id: If9ffe9547619f54c63b491e9b5ac5a8d91c35644
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1084274
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: David Worsham <dworsham@google.com>