[power,test] Disentangle PowerSampler and PowerMetricsProcessor
Initially, test code needed to determine at runtime whether to use
PowerMetricsProcessor or a stub. This was implemented by providing an
affordance to ask a PowerSampler implementation to give back the right
metrics processor to use.

There are no more callers of the old API, so remove it.

Bug: None
Test: fx test power_test_utils_test
Change-Id: I1015ca60d55bbd51daa6c18c88bdbd2770714c58
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093036
Reviewed-by: Gwen Mittertreiner <gmtr@google.com>
Commit-Queue: Chris Masone <cmasone@google.com>