[fxfs] Get or insert keys to use None instead of NotFound errors.
Lets the caller decide how to handle a missing encryption key. In the
case of prefetch-keys it is best effort and the object may be deleted by
the time the prefetch arrives. In most other cases where a key is
expected they can turn it into an Inconsistent error.

Notably this handling change will prevent the error logged failure in
`UnwrapResult::Set` when the key is missing from the tree, as only the
caller can decide if this is an issue.

Bug: b/356626753
Change-Id: I8694ee99d76bc97ffa031532959e60fe63414379
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093072
Commit-Queue: Martin Lindsay <mlindsay@google.com>
Reviewed-by: James Sullivan <jfsulliv@google.com>