[routing] simplify AllowedSourcesBuilder.framework
The optional framework field in AllowedSourcesBuilder held a closure to
convert a Name to an InternalCapability. In practice this closure only
ever converted a name into the same InternalCapability as the
CapabilityTypeName declared on the AllowedSourcesBuilder.

To simplify things this closure has been replaced with a boolean field.
When set to true, the framework_sources function will return an
InternalCapability matching the CapabilityNameType. There is no behavior
change due to this.

Change-Id: If032979a863defb52734f9842c3b8b61d17223a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1091933
Reviewed-by: David Gilhooley <dgilhooley@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>