[audio][test] VirtualAudio VMO rights
Update virtual_audio driver (Composite, Dai, StreamConfig)
to apply the appropriate handle rights to the VMO that it
returns in response to `RingBuffer/GetVmo`.

In the process of doing this, also fixed the Composite
implementation of `WatchTopology` (which tests use to
determine whether ZX_RIGHT_WRITE is required) so that
driver signalprocessing state is correctly reset when the
Composite protocol channel is closed.

Test: audio_driver_tests
Test: audio-core-api-pipeline-tests

Change-Id: I258879142a355cfc1e34dab1ed96a57ca7ca7bb2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093753
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>