[rbe][cxx] Workaround for newer clang toolchain
Newer versions of clang drive the compiler in a manner that
is sensitive to the existence of target libdirs.
re-client isn't aware of this yet, and only includes the correct headers
as remote inputs, without any linker directories.

This temporarily works around the issue by adding one file per critical
target libdir as a remote input to force those directories to
exist in the remote environment.  Since this wrapper is only used for
compilation and not linking, these extra remote inputs are not actually
used.

Bug: 354016617
Change-Id: Ie7a4b5cc974ec5213c452ebdfce0033448bf8e0c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1092316
Commit-Queue: David Fang <fangism@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>