[kernel] Modify UpdatePlatformTimer to choose the soonest deadline
Adds back the UpdatePlatformTimer method to TimerQueue, but have the new
version take no arguments and set the platform's oneshot timer to the
soonest deadline among:
1. The preemption deadline.
2. The next monotonic timer's scheduled time.
3. The next boot timer's scheduled time.

This is currently only used by the `TimerQueue::Tick` and `Timer::Cancel`
functions, as other operations can be more efficient and only compute the
deadline on a particular timeline. For example, when setting a monotonic
timer, we do not need to recompute the raw ticks deadline of the head of
the boot timer list, so we need only call UpdatePlatformTimerMono.

In the future, this method will be used by the IdlePowerThread to update
the platform timer after the monotonic timeline is paused during
suspension.

Bug: 341785588
Change-Id: Ib2d50d92ff102fd618f6d995ec88a26e29be373c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1092478
Reviewed-by: John Grossman <johngro@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>