[kernel] Convert IdlePowerThread to boot time
Converts the idle power thread to use boot time when transitioning the
system into suspend. This required changing some tests as well.

Note: This change does not pause the monotonic clock on suspension, so
the boot and monotonic timelines will never diverge. Thus, this change
should be a functional no-op. It is being made solely to minimize the
amount of code that must be checked in behind a feature flag.

Bug: 341785588
Change-Id: I41630fbe5cae864b2edf191757bc4f39b5161a73
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094013
Reviewed-by: Corey Tabaka <eieio@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Anirudh Mathukumilli <rudymathu@google.com>