[superproject] Roll llvm-project/libc [libc] enable most of the entrypoints on aarch64 (#101797)
This is a non-feature change that enables most of the entrypoints for
aarch64 based runtime builds. It fixes an additional problem that some
compiler-rt targets are not defined at the time of dependency checking
thus leading to false-negatives.

GitOrigin-RevId: 5fdcef221fd3ee1b7287023c6d488613430fcd4a
Original-Revision: 10597217ee75f9c6636d7af729966a1fd859a2d6
Roller-URL: https://ci.chromium.org/b/8740583805707113921
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I10e0f61b494a76d5f80d9242a00089e2597470e9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094212