[kernel][vm] Page request status is first page
Makes the signalled status of a page request be the status of the first
page, instead of any failed portion. This allows for the case where the
first pages in the request are supplied successfully and later pages
are failed to be correctly handled.

Run-All-Tests: True
Change-Id: I047c109d79c0a6b8094b67edd91813096de2750e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1091097
Commit-Queue: Adrian Danis <adanis@google.com>
Reviewed-by: David Worsham <dworsham@google.com>