[netstack3] Lookup Multicast Routes during IP RX
When handling a received IP Multicast packet, query the routing table
and determine which targets the packet should be forwarded to, if any.
For conformance with Linux/NS2, attempt to forward multicast packets
regardless of whether the host has membership in the multicast group.

For now, the packet isn't actually forwarded.

Bug: 353329136
Change-Id: I440d576d1b7c4c59599a15b4cd0a8b61a39d50dd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1092872
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>