[netstack3] MulticastForwardingDeviceContext
Move the check for whether multicast forwarding is enabled on a device
to its own context trait. This moves in the direction of narrowly scoped
context traits, which reduces the amount of boilerplate required in
unit tests.

This is the follow up to a comment left on https://fxrev.dev/1089037

Bug: 353329136
Change-Id: I1986542b8413fffb1af424b1798dd5aac860fc8c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093173
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>