[bazel-sdk] check for extra resources in package checks
The package_checker_tests were only looking for expected manifets and
did not check ot see if manifests that should not be included were being
included. This allowed some of the generated cml files to be included in
the package under the wrong name. This was not breaking the packages
since the file was also included under the correct name but it caused
confusion and was just not correct.

BUG: b/356932372
Change-Id: I34d1a343fe631935caf13b92a49d97cf3ef105e2
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094235
Reviewed-by: Darren Chan <chandarren@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>
Reviewed-by: Jiaming Li <lijiaming@google.com>