[bt-host] fix test_component names in build files
The bazel fuchsia_unittest_package is bundling the component manifests
twice. One of the manifests is named after the component and one is
named after the generated file even though they are both the same source
file. We are cleaning up this logic but we need to first update the
usages of the incorrect name.

BUG: b/356932372
Change-Id: I6c06376a7f7daaa60fe2c4a6b0a47df7e783f9de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094236
Reviewed-by: Darren Chan <chandarren@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>