[starnix][selinux] Add fs_node_set/getsecurity() hooks
fs_node_setsecurity() allows the security module to update state
associated with the in-memory representation of a file system node
whenever its value is modified.

fs_node_getsecurity() allows the security module to answer queries
for the security attribute of any file node by returning the
Security Context that it is labelled with. This allows userspace
tooling to query the security labels on files, if necessary,
regardless of whether the file system actually supports xattrs,
or was mounted with fs_use_xattr labelling.

If SELinux is not enabled, or some other "security.*" is accessed
then the hooks delegate to the underlying FsNodeOps.

Bug: b/351195217
Change-Id: I626b6da712600aed65951487a89ea9e457525e25
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1092232
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>