[fuchsia-controller] Adapter code for async/non-async Python.
This creates a convenience mixin for running async Python in a non-async
context. This is mainly intended to be used for adapting test frameworks
like Mobly, which don't support running async methods, to be able to run
easily readable code that has well-defined async behavior.

Each instance of an object here contains an async loop that is around
for the lifetime of the class. Each function call using the
`@asyncmethod` decorator runs through this loop.

Bug: b/357147229
Change-Id: I9ceb830065d619db27e5916ad2f7b34b9a7e990c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094012
Reviewed-by: Sam Balana <sbalana@google.com>
Reviewed-by: Pavan Kumar Juturu <jpavankumar@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>
Commit-Queue: Andrew Davies <awdavies@google.com>