[ffx][playground] Simplify parser
The usual mode for error handling in parsers is just to try to skip past
the error and resume parsing, with the parse always returning a tree but
the tree sometimes containing error information.

This might seem like it poses a problem with parser combinators. If you
have a parser for `declarations` and a parser for `expressions`, and you
try to parse either with `alt((declarations, expressions))`, if
declarations *always* succeeds, even given nonsensical input, you'll
never parse an expression no matter what.

Playground previously tried to get around this with a complicated error
recovery mechanism that would run the parse multiple times on error,
allowing continuation in various places until a parse with minimal
errors succeeded.

It turns out that the best practice is instead to just be aware of the
above complication and carefully design your grammar around it, e.g.
`alt((declarations_no_errors, expressions, declarations_maybe_errors))`.

This lets us cut a huge swathe of complexity out of the parser code and
the grammar adjustments aren't as major as you might expect.

Change-Id: I0ac8f6c60de8d5154c43196f64bc04813fe90235
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1092556
Fuchsia-Auto-Submit: Casey Dahlin <sadmac@google.com>
Commit-Queue: Casey Dahlin <sadmac@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>