[audio] audio_driver_tests validates VMO rights
The audio driver test suite now validates that the VMO
handles returned from an audio driver has the required
rights. In order to do this, the test must reason about
whether the ring buffer is 'outgoing' or 'incoming'.
This is trivial for Dai and StreamConfig drivers (the
driver declares the `is_input` bool in its properties)
but for Composite drivers it must understand the ring
buffer element's place in the signalprocessing topology.

Test: audio_driver_tests -- --admin

Change-Id: I257195f01d08bf2d0abfbeef9e597589439ff962
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093754
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>