[kernel][phys] Introduce formal end of hand-off
This change introduces a formal end to the hand-off phase via the
function `EndHandoff()`. It is currently expected to be called in
userboot, unsets gPhysHandoff, returns the remaining hand-off data left
to be consumed (e.g., the ZBI in the physmap), and populates its subset
of userboot's VMO handles.

Previously, the handle population logic was built into the
instrumentation VMO preparation code, as that seemed a convenient spot
to put it at the time and there were similarities. This is now naturally
moved out to a hand-off-specific site.

Moreover, as of this change, <phys/handoff.h> is now evaluated within
the userboot toolchain.

Bug: 357155617
Change-Id: I3036e4ccff1646a5ca5c47d2acc4a3944bcc3daa
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094015
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>