[kernel][mmu] Remove DropPAddrs
The MappingCursor::DropPAddrs method exists to allow a MappingCursor to
switch to a mode where it only tracks virtual addresses, allowing it to
consume a non-page sized portion for error purposes. Consuming such a
large portion of virtual address is excessive for the outcome needed,
which is to ensure that the table is checked for emptiness, and a
single PAGE_SIZE slot being consumed is sufficient.

Removing the DropPAddrs method will allow for splitting the
MappingCursor into type safe variants that are either always tracking
paddrs, or always tracking just a virtual range.

Change-Id: I4033cdff925d0523ec7a75a05878591d0837cf9b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093473
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>