[kernel][mmu] Separate VirtualAddressCursor
Splits the mapping cursor into a MappingCursor that always tracks
physical addresses to map, and a VirtualAddressCursor that only tracks
a virtual range.

This provides a safer set object as previously there were methods on
the MappingCursor that were only valid to be called if it was in the
correct mode and this had to be asserted. Now these methods are just
on their respective specific implementation and mistakes cannot be
made.

Change-Id: I0daa9a49a749729e488cfd2bae57917c98e51302
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1093474
Reviewed-by: Travis Geiselbrecht <travisg@google.com>
Reviewed-by: David Worsham <dworsham@google.com>
Fuchsia-Auto-Submit: Adrian Danis <adanis@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>