[aml-rawnand] Wait for command queue on reads
Wait for the command queue to be empty when reading before checking the
status bits.

The nand cycle time for the Micron chip is 24ns and the page size is
1024. This results in sleeping for 24.5us. Waking up, checking the bit,
and sleeping again can take ~5us. 4k is read from the device at a time
which should result in sleeping 4 times and taking roughly 118us. From
looking at traces, a 5th sleep is frequently required. 5 sleeps * 5us =
25us of CPU time spent per 4k read. Blobfs typically reads in 6 pages at
once, 25us * 6 = 150us. That's half of the nand-worker's deadline
scheduler budget just sleeping.

Multiplying the number of ecc_pages into the sleep would only sleep for
98us and more polling with a shorter interval would be necessary.

Testing different sleep durations shows that 130us works almost all the
time. 130us is also the command queue flush initial sleep.

This change goes back to the previous implementation which waited for
the command queue to be empty. That single 130us sleep worked 99.99% of
the time. To account for the 0.01%, the status bits are still checked
with a short polling interval.

Bug: b/350020539
Change-Id: If870e20c4b62d12a339e033271d4257d80a313be
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1092194
Reviewed-by: Braden Kell <bradenkell@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>
Reviewed-by: Chris Suter <csuter@google.com>