[display][coordinator] Always bind the coordinator ServerEnd to Client.
Previously in some coordinator tests, the coordinator server-end channel
was not bound to a `Client` instance but kept as a standalone ServerEnd
and stored in the DisplayControllerBindingState type.

This change removes the test-only constructors for Client and
ClientProxy that store the ServerEnd,  and replaces them with
`BindForTesting() / InitForTesting()` methods that bind the ServerEnd
to `Client` instead. This unifies the client binding logic and makes it
possible for us to remove the `DisplayControllerBindingState` type used
to emit events.

Test: no production code affected.
Test: display-coordinator-unittest
Bug: 42075581
Change-Id: Idb9034e0f33fccbda352de779cd386254d8b73cb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1081902
Reviewed-by: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>