Reland [ffx][emu] Refactoring disk hashing to emu_instance
> Reason for revert: Broke mac builders
> Failed builders: go/bbid/8740400844815955649, go/bbid/8740397533543846081

The original was reverted because of a build error on mac. This has been
fixed.

This refactors the API used to hash the disk images to be
in emulator_instance so it is easier to use this when the
configuration is using EFI vs. ZBI to start the emulator.

> Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094269
> Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
> Reviewed-by: Andrew Davies <awdavies@google.com>

Change-Id: I296849f1033b5194f65f53f778b209435471ba66
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094835
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Andrew Davies <awdavies@google.com>