[bazel-sdk] decouple autogenerated cc_tests
This CL removes the autogenerated cc_tests from the build graph when
they are not needed. We were always placing the generated components in
the build graph because they were added to the FuchsiaUnitTestComponentInfo
provider. This made it so the aspects that collect resources would pick
up the manifests incorrectly.

Additionally, we updated the fuchsia_component_manifest logic to use the
component_name for the output name if provided which matches the in-tree
logic.

BUG: b/356932372
Change-Id: Ife2b944ebfb6eebe370b644425675b24c65ceca9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094237
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: Darren Chan <chandarren@google.com>
Commit-Queue: Chase Latta <chaselatta@google.com>