[windowed-stats] Implement serialization
Serialize all the information needed to deserialize the data from a
TimeMatrix. Adjust the allowable types for a TimeMatrix to match
what's in the design document, as otherwise we don't have a type
descriptor to include in the serialized buffer.

Update the `interpolate` method to no longer return a serialized buffer.
This way, we provide an option to update the TimeMatrix without having
to incur the cost of serialization.

Bug: 350797382
Test: Added unit test
Change-Id: I9138f46a3105d4ceb0df5789088515a3820a10c9
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1084873
Fuchsia-Auto-Submit: Kiet Tran <kiettran@google.com>
Reviewed-by: Devon H. O'Dell <dhobsd@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>