[abi-compat] 25% speedup through conversion refactor
This inclues some changes to how we convert from the IR types to the
comparison types. The use of FlyStr is significantly reduced because it
was showing up high in traces and it has properies (like thread safety)
that don't matter here. Instead of allocating and copying Vecs to track
state and nesting as we walk through types, we use references to values
on caller's stacks. These paths are still turned into Vecs with Strings
when storing them on the converted values, and there's more room for
improvements there.

Change-Id: Id62113fa5ab5e9e870aad33a7e6c1aa965e66429
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094064
Reviewed-by: Mitchell Kember <mkember@google.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>