[abi-compat] More optimizations
This changes the compare::Path type to be much simpler, just carrying
the version and a string representing the path (rather than a Vec of
path elements). The code to generate path strings moves to
convert::Context and is more efficient - calculating the length of the
string before allocating the space all at once and filling it in.

This cleans up the compare::ProblemPaths type to be simpler and hangs
onto values used for display and comparison rather than fetching or
calculating them.

This is an incremental 33% improvement. Overall relative to baseline
this is a 56% improvement.

Change-Id: Ie08f885061712d4e2aa890030e19dd771813fe41
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1096172
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Mitchell Kember <mkember@google.com>