[abi-compat] Clean up & optimize error reporting
This includes a bunch of simplifications and optimizations in error
reporting. Originally the API levels and paths of conflicting
declarations made up part of the error message (CompatibilityProblem)
type, but as the tool developed the API levels, ended up in error
messages too, and we only ended up referring to common paths. So this
removes API levels and individual paths from CompatibilityProblem and
also the types we use for tests.

Change-Id: I5c806cdefa31ecd75183e192e39dfe6524102c07
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1096173
Reviewed-by: Mitchell Kember <mkember@google.com>
Commit-Queue: Ian McKellar <ianloic@google.com>