[libc] Temporarily disable UBSan on __dls3 for riscv64
The bringup.riscv64-asan builder was crashing very early during bringup
with a stack overflow in the next clang roll.
-Rpass-analysis=stack-frame-layout showed a bump in stack usage for
__dls3 (almost 56 extra bytes). Disabling UBSan for this particular
function on RISCV results in the same stack usage before and after the
clang roll. It's likely that upstream changes to UBSan resulted in extra
stack usage. We've experienced this type of issue before regarding UBSan
and bringup.riscv64-asan. We have plans to address this in the long run,
but for now we can just suppress UBSan for just this one function to
unblock the clang roll.

Bug: 42080808, 357634850
Change-Id: I6ddba9b148d13599de4a1aec706d2855722ca605
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1097105
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Leonard Chan <leonardchan@google.com>