[ffx][emu]Changing --dry-run to run after staging
This changes the emulator setup flow to stage the emulator instance
before printing the command line when the `--dry-run`  option is
provided.

This is being done to make the `--dry-run` output match the actual
command line that would be run to start the emulator.

Change-Id: I16be631be05beb11e8cbd43ea115324c827087eb
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1097796
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Steven Grady <slgrady@google.com>