[netstack3/conntrack/tcp] Move self in State::update
...and return a new state to replace the old one.

Most of the upcoming helpers use this pattern, and pushing it all the
way up to the highest level means it can be used in more places.

As part of this, we add a "Closed" terminal state, which is used to
signal that a connection has been closed and should be removed from the
conntrack table.

With all of this simplification, the return value of the update function
has become (State, bool), where the boolean is whether the segment is
valid.

Bug: 328064736
Change-Id: Icd87dc9c921ab19b82d24d0b70f420d2526b8204
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1092877
Commit-Queue: Peter DeLong <pdelong@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>