[microbenchmarks] Reduce variation on VIM3 by setting big core affinity
Attempts to detect the vim3 board and set the affinity of the
benchmark thread to the big CPUs, this prevents them from migrating
between the big and little CPUs creating large amounts of benchmark
variation.

This is intended to be a temporary change until the scheduler can
properly prefer scheduling threads on big cores over little cores when
running the benchmarks.

Bug: 42050716
Change-Id: I7e5b3e647c3a0d952031b8e65b9fedbdab4593a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1087293
Reviewed-by: Mark Seaborn <mseaborn@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>