[superproject] Roll scudo [scudo] Added test fixture for cache tests. (#102230)
The test fixture simplifies some of the logic for allocations and
mmap-based allocations
are separated from the cache to allow for more direct cache tests.
Additionally, a couple
of end to end tests for the cache and the LRU algorithm are added.

GitOrigin-RevId: 87d9b29e2996edff083375d22495473d1b3f09e7
Original-Revision: 72353a410ab866d268bd27d00f8f63f2c9d4d56b
Roller-URL: https://ci.chromium.org/b/8740132063389131633
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: I0af24bd4ea516ffbb88ab5d4b51dbbe998e3c87e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1098017