[kernel] Select PMM arenas algorithmically
This change updates PMM initialization to be driven by the normalized
accounting of hand-off-described memory, running that through the new
selection algorithm in <vm/phys/arena.h>.

See the commit message of fxrev.dev/1074514 for the limited arena
selection differences for supported boards.

In the interest of incrementality, we continue to use bootreserve to
manage reserved regions. This will be removed in a follow-up.

Bug: 347766366
Change-Id: Ia079ac47282fff5550314800c1fabad5f085b8de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1076373
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Adrian Danis <adanis@google.com>
Commit-Queue: Joshua Seaton <joshuaseaton@google.com>
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>