[netstack3/conntrack/tcp] Closing
Like with Established, this is a thin wrapper around
do_established_update.

There are two things we're not planning on handling (for now).

The first is that we aren't going to handle data segments after a FIN is
sent. See the comment at the top of Closing::update for more details on
this. Since it's not handled by gVisor or Linux, it doesn't seem
important for us to, and shouldn't affect correctness.

Second, we are not implementing a TimeWait state. That state is not
required for correctness, as is further described by the comment at the
end of Closing::update where the transition would occur.

Bug: 328064736
Change-Id: I37613e8e66d10bfa98062b139b8ffe5fe387adb3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1091353
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>