[netstack3/conntrack] Fix establishment tracking
We're attempting to mimic Linux behavior, but had done this improperly
before. We had assumed that "ASSURED" meant the same as "ESTABLISHED",
but Linux actually uses a three-state tracking system for this.

Furthermore, this is unrelated to the underlying protocol, so TCP
"ESTABLISHED" (handshake successfully completed) is totally different
from conntrack "ESTABLISHED" (packets seen in original, reply, and
original direction).

While this doesn't really matter until we have a stateful firewall, it
was easier to get done while we're already working on conntrack.

Fixed: 355699112
Change-Id: I8de1ce1b52308ee31fb6ccf7a13dfaef028985a1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1091417
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>