[routing] every RoutingError variant has a moniker
In order to preserve UX in scrutiny, every routing error variant should
include a moniker indicating where the routing error happened. This is
necessary because today this information is guaranteed by the presence
of list of route segments in the scrutiny output, and in a future change
those route segments will be removed. This error type thus must carry
all the necessary information for diagnosing and resolving the issue it
signifies.

Change-Id: I9250b3923a213c7cf1db514f6b664a8a4dac77b7
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1094672
Reviewed-by: Gary Bressler <geb@google.com>
Commit-Queue: Claire Gonyeo <cgonyeo@google.com>