[netstack3] Fix dangling neighbor timer
Currently, when a neighbor entry is created in the INCOMPLETE state, the
timer to retransmit a multicast neighbor solicitation is unconditionally
scheduled, and only then is the packet to be sent out queued. However,
serialization of the pending packet is fallible, and if it fails, the
timer that was scheduled remains in the timer heap; if a future timer is
scheduled for this neighbor, it trips an assertion that expects no other
timers to be scheduled.

Fix this by only scheduling the timer once serialization of the pending
packet was successful.

Fixed: 357115127
Test: netstack3-core-ip-test
Change-Id: I7eb8b177c5cb6ded988d4bce5500493bf18dc693
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1097992
Reviewed-by: Evan Wildenhain <wildenhain@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>