[starnix][selinux] Revise logic for FsNode security label choice
fs_node_effective_sid() always returns a SID, suitable for use
to access-check the file. It uses the SID attached to the FsNode
if any, otherwise delegating to fs_node_resolve_security() to
attach one.

fs_node_resolve_security() chooses a SID based on the file-system
labeling scheme and mount options, including attempting to resolve
labels stored in extended attributes, if appropriate. This process
is fallible but if a SID is resolved then it is attached to the
FsNode and returned to the caller.

Bug: b/355587237, b/334094811, b/349800754
Change-Id: I44912711228bce3d9a9f5e992e9e7a7b556c0152
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1095872
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Fuchsia-Auto-Submit: Wez <wez@google.com>
Commit-Queue: Wez <wez@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>