[sdmmc] Update to instance name
Update the power config so that it doesn't need to be edited.
This driver always wants to use the 'default' instance for the
power element so use that instead of the parent name which changes.

Fixed: b/356128309
Change-Id: I62cd690b7526558f6feaa9848c452f61c43e0e09
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1089413
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>
Fuchsia-Auto-Submit: David Gilhooley <dgilhooley@google.com>