[starnix][task] Check for X during exec()
Previously, we were checking for the R throughout the interpreter chain
and then checking again for X at the end of the interpreter chain. We
now check for X (and not R) throughout the interpreter chain.

We still fail SuidTest.SuidBinaryBecomesRoot, but we get much further.
For example, we're now able to actually exec the test binary.

Bug: b/329142824
Test: SuidTest.SuidBinaryBecomesRoot
Test: InterpreterScriptTest.BinaryNoExecute
Test: prctl06
Change-Id: I674ad39fbe3cc3f73dd784cb6e4048cbc71c3a78
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1097555
Reviewed-by: Theodore Dubois <tbodt@google.com>
Fuchsia-Auto-Submit: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>