[vfs][cpp] Bind SynchronousVfs connections under lock
Similar to ManagedVfs, ensure we call Bind() while still holding the VFS
lock. Not holding the lock was leading to flakes on debug builders when
checking that the connection itself was still owned by the container,
since there was a small window where the VFS could be shutdown before
Bind() was called.

This also ensures that the BindImpl() functions in each connection type
do not dereference nullptr when accessing the dispatcher associated with
the VFS.

Fixed: 358451182
Test: fx test fs-connection-test
Multiply: fs-connection-test
Change-Id: Ib76cf88ce4ea877dda8ff4bdba1310db7fdfe4c1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1070953
Reviewed-by: James Sullivan <jfsulliv@google.com>
Commit-Queue: Brandon Castellano <bcastell@google.com>