[storage] Add headers to build targets
Make sure all storage headers belong to a build target. Nothing ensures
that included headers are part of target. But if a header is part of a
target then GN will make sure that the header's target is depended on
by the includer's target. This ensures that visibility rules are
enforced.

Change-Id: If1cb09328758b7f27006fdb137361e82a67a3a06
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1097716
Reviewed-by: Martin Lindsay <mlindsay@google.com>
Commit-Queue: Chris Drouillard <cdrllrd@google.com>