[build] Don't use -Wc++20-compat
This switch is meant to be used in C++ 17 mode to catch things
that would need to change for 20.  However, we were using it
instead only in 20 mode, which is not useful.  There are too many
warnings in 17 mode for things that are already conditionalized
appropriately for 20 mode so it's not useful at all.

Bug: 42064981
Change-Id: I1ae14a65146d53e336b07b7deec05b5bdf90de27
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1068733
Commit-Queue: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Petr Hosek <phosek@google.com>