[driver][test-realm] realm-builder remove shard
This removes the process launcher offer to the driver test realm as it
is not used after the dfv2 migration. This also allows us to remove the
shard that we had which only had this entry in it.

Since the shard itself is accessed in vendor repos we will keep it in
this patch, then when those are removed we can delete the shard.

Change-Id: I89fa27ea9cd30260dd8fea2542a480c8fc691191
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1098075
Reviewed-by: Suraj Malhotra <surajmalhotra@google.com>
Fuchsia-Auto-Submit: Novin Changizi <novinc@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Aaron Wood <aaronwood@google.com>